WE-339 dismiss snackbar after 3 seconds
